+++ tests/Fixtures/Core/Extension/StubExtension.php
@@ -35,8 +35,8 @@ final readonly class StubExtension implements ExtensionInterface
     /**
      * @param non-empty-string                                                                     $identifier
      * @param array<string, mixed>                                                                 $settings
-     * @param array<non-empty-string, class-string<JsonRpcRequest<non-empty-string>>>              $requests
-     * @param array<non-empty-string, class-string<JsonRpcNotification<non-empty-string>>>         $notifications
+     * @param list<class-string<JsonRpcRequest<non-empty-string>>>                                 $requests
+     * @param list<class-string<JsonRpcNotification<non-empty-string>>>                            $notifications
      * @param array<non-empty-string, RequestHandlerInterface<non-empty-string, Result, TContext>> $requestHandlers
      * @param array<non-empty-string, NotificationHandlerInterface<non-empty-string>>              $notificationHandlers
      */