+++ tests/Fixtures/Server/Extension/StubServerExtension.php
@@ -31,8 +31,8 @@ final readonly class StubServerExtension implements ServerExtensionInterface
     /**
      * @param non-empty-string                                                                          $identifier
      * @param array<string, mixed>                                                                      $settings
-     * @param array<non-empty-string, class-string<JsonRpcRequest<non-empty-string>>>                   $requests
-     * @param array<non-empty-string, class-string<JsonRpcNotification<non-empty-string>>>              $notifications
+     * @param list<class-string<JsonRpcRequest<non-empty-string>>>                                      $requests
+     * @param list<class-string<JsonRpcNotification<non-empty-string>>>                                 $notifications
      * @param array<non-empty-string, RequestHandlerInterface<non-empty-string, Result, ServerContext>> $requestHandlers
      * @param array<non-empty-string, NotificationHandlerInterface<non-empty-string>>                   $notificationHandlers
      */